Misc fixes for review
